+++ src/sentry/static/sentry/app/components/events/interfaces/stacktraceLink.tsx
@@ -1,12 +1,9 @@
 import React from 'react';
-import Modal from 'react-bootstrap/lib/Modal';
 import styled from '@emotion/styled';
 
 import AsyncComponent from 'app/components/asyncComponent';
 import Button from 'app/components/button';
-import ButtonBar from 'app/components/buttonBar';
-import {t, tct} from 'app/locale';
-import space from 'app/styles/space';
+import {t} from 'app/locale';
 import {
   Event,
   Frame,
@@ -18,9 +15,9 @@ import {
 import {getIntegrationIcon, trackIntegrationEvent} from 'app/utils/integrationUtil';
 import withOrganization from 'app/utils/withOrganization';
 import withProjects from 'app/utils/withProjects';
-import InputField from 'app/views/settings/components/forms/inputField';
 
 import {OpenInContainer, OpenInLink, OpenInName} from './openInContextLine';
+import StacktraceLinkModal from './stacktraceLinkModal';
 
 type Props = AsyncComponent['props'] & {
   frame: Frame;
@@ -40,8 +37,6 @@ type StacktraceResultItem = {
 
 type State = AsyncComponent['state'] & {
   match: StacktraceResultItem;
-  showModal: boolean;
-  sourceCodeInput: string;
 };
 
 class StacktraceLink extends AsyncComponent<Props, State> {
@@ -103,24 +98,6 @@ class StacktraceLink extends AsyncComponent<Props, State> {
     };
   }
 
-  openModal() {
-    this.setState({
-      showModal: true,
-    });
-  }
-
-  closeModal() {
-    this.setState({
-      showModal: false,
-    });
-  }
-
-  onHandleChange(value: string) {
-    this.setState({
-      sourceCodeInput: value,
-    });
-  }
-
   onOpenLink() {
     const provider = this.config?.provider;
     if (provider) {
@@ -154,6 +131,10 @@ class StacktraceLink extends AsyncComponent<Props, State> {
     }
   }
 
+  onClose() {
+    this.reloadData();
+  }
+
   // let the ErrorBoundary handle errors by raising it
   renderError(): React.ReactNode {
     throw new Error('Error loading endpoints');
@@ -163,84 +144,25 @@ class StacktraceLink extends AsyncComponent<Props, State> {
     //TODO: Add loading
     return null;
   }
+
   renderNoMatch() {
-    const {showModal, sourceCodeInput} = this.state;
     const {organization} = this.props;
     const filename = this.props.frame.filename;
-    const baseUrl = `/settings/${organization.slug}/integrations`;
 
-    if (this.integrations) {
+    if (this.project && this.integrations.length > 0 && filename) {
       return (
-        <CodeMappingButtonContainer columnQuantity={2}>
-          {t('Enable source code stack trace linking by setting up a code mapping.')}
-          <Button onClick={() => this.openModal()} size="xsmall">
-            {t('Setup Code Mapping')}
-          </Button>
-          <Modal
-            show={showModal}
-            onHide={() => this.closeModal()}
-            enforceFocus={false}
-            backdrop="static"
-            animation={false}
-          >
-            <Modal.Header closeButton>{t('Code Mapping Setup')}</Modal.Header>
-            <Modal.Body>
-              <ModalContainer>
-                <div>
-                  <h6>{t('Quick Setup')}</h6>
-                  {tct(
-                    'Enter in your source code url that corresponsds to stack trace filename [filename]. We will create a code mapping with this information.',
-                    {
-                      filename: <code>{filename}</code>,
-                    }
-                  )}
-                </div>
-                <SourceCodeInput>
-                  <StyledInputField
-                    inline={false}
-                    flexibleControlStateSize
-                    stacked
-                    name="source-code-input"
-                    type="text"
-                    value={sourceCodeInput}
-                    onChange={val => this.onHandleChange(val)}
-                    placeholder={t(
-                      `https://github.com/octocat/Hello-World/blob/master/${filename}`
-                    )}
-                  />
-                  <ButtonBar>
-                    <Button type="button" onClick={() => {}}>
-                      {t('Submit')}
-                    </Button>
-                  </ButtonBar>
-                </SourceCodeInput>
-                <div>
-                  <h6>{t('Manual Setup')}</h6>
-                  {t(
-                    'To set up a code mapping manually, select which of your following integrations you want to set up the mapping for:'
-                  )}
-                </div>
-                <ManualSetup>
-                  {this.integrations.map(integration => (
-                    <Button
-                      key={integration.id}
-                      type="button"
-                      to={`${baseUrl}/${integration.provider.key}/${integration.id}/?tab=codeMappings`}
-                    >
-                      {getIntegrationIcon(integration.provider.key)}
-                      <IntegrationName>{integration.name}</IntegrationName>
-                    </Button>
-                  ))}
-                </ManualSetup>
-              </ModalContainer>
-            </Modal.Body>
-            <Modal.Footer></Modal.Footer>
-          </Modal>
-        </CodeMappingButtonContainer>
+        <StacktraceLinkModal
+          filename={filename}
+          project={this.project}
+          organization={organization}
+          integrations={this.integrations}
+          onClose={() => this.onClose()}
+        />
       );
     }
     return null;
   }
+
   renderMatchNoUrl() {
     const {config} = this.match;
     const {organization} = this.props;
@@ -284,29 +206,4 @@ export {StacktraceLink};
 
 export const CodeMappingButtonContainer = styled(OpenInContainer)`
   justify-content: space-between;
-`;
-
-const SourceCodeInput = styled('div')`
-  display: grid;
-  grid-template-columns: 5fr 1fr;
-  grid-gap: ${space(1)};
-`;
-
-const ManualSetup = styled('div')`
-  display: grid;
-  grid-gap: ${space(1)};
-  justify-items: center;
-`;
-
-const ModalContainer = styled('div')`
-  display: grid;
-  grid-gap: ${space(3)};
-`;
-
-const StyledInputField = styled(InputField)`
-  padding: 0px;
-`;
-
-const IntegrationName = styled('p')`
-  padding-left: 10px;
 `;
